test: add two more test cases for extractors
This was prompted by the coverage report: it was complaining that the
lambdas defined for the first Amazon and location extractors were not
executed. Cryptically, at that. The message was:
`line 107 didn't jump to the function exit`
Line 107 has the `e += [ E(..., P(lambda m: ...)), ... ] and it was that
innermost lambda that did not execute during `extractors_test.py`.

Funnily enough, it's completely A-OK to have other extractors there that
didn't match, as long as they don't have lambdas - they get defined, all
code is being ran, and from the coverage PoV, all is fine :D